# quell-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Quell AI Backend (derived from routes)
  version: 1.0.0
  description: |
    OpenAPI specification for the Quell AI backend API. Endpoints cover authentication,
    store management, Shopify OAuth, chat webhook, inbox, support tickets, appearance,
    analytics, and admin metrics.
servers:
  - url: https://histioid-imogene-spinally.ngrok-free.dev
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
    TokenResponse:
      type: object
      properties:
        success: { type: boolean }
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    Store:
      type: object
      properties:
        id: { type: integer }
        shop_domain: { type: string }
        store_id: { type: string }
        widget_token: { type: string }
        product_count: { type: integer }
        status: { type: string }
        last_sync_at: { type: string, format: date-time }
        installed_at: { type: string, format: date-time }
...
security:
  - BearerAuth: []
paths:
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user and link a store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
                name: { type: string }
                shopToken: { type: string }
              required: [email, password, name, shopToken]
      responses:
        '200':
          description: Created with token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        '400':
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        '401':
          description: Invalid credentials
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user and their stores
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user and stores
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  stores:
                    type: array
                    items: { $ref: "#/components/schemas/Store" }
...
  /api/customer-tickets:
    get:
      tags: [CustomerTickets]
      summary: List customer tickets for a seller dashboard
      parameters:
        - in: query
          name: store_url
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: customer_id
          schema: { type: string }
      responses:
        '200':
          description: List of customer tickets

    post:
      tags: [CustomerTickets]
      summary: Create a ticket from the chatbot on behalf of a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                store: { type: string }
                customerId: { type: string }
                name: { type: string }
                email: { type: string }
                subject: { type: string }
                description: { type: string }
                orderId: { type: string }
                sessionId: { type: string }
                cartID: { type: string }
              required: [store, customerId, subject, description]
      responses:
        '201':
          description: Created customer ticket
  /api/customer-tickets/{id}:
    get:
      tags: [CustomerTickets]
      summary: Get a single customer ticket by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Customer ticket

    patch:
      tags: [CustomerTickets]
      summary: Update customer ticket fields
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                priority: { type: string }
                subject: { type: string }
                description: { type: string }
      responses:
        '200':
          description: Updated
  /api/appearance/{storeUrl}:
    get:
      tags: [Appearance]
      summary: Get chatbot appearance and branding settings for a store
      parameters:
        - in: path
          name: storeUrl
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Appearance object

    put:
      tags: [Appearance]
      summary: Save or update chatbot appearance settings
      parameters:
        - in: path
          name: storeUrl
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                primary_color: { type: string }
                header_title: { type: string }
                welcome_message: { type: string }
                conversation_starters:
                  type: array
                  items: { type: string }
                button_bg_color: { type: string }
                button_text_color: { type: string }
                button_shape: { type: string }
                button_position: { type: string }
                logo_url: { type: string }
                show_logo: { type: boolean }
      responses:
        '200':
          description: Saved
...